CUDA-specific TcExecutor moves to core/cuda
